add api.modifyDefaultConfig
